+++ model/src/main/java/cn/qihangerp/model/request/StockInItem.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class StockInItem {
     private Long id;
-    private Long intoQuantity;
+    private Integer intoQuantity;
     private Long positionId;
     private String positionNum;
 }
